Addresses MAC OS numpy singularity handling
numpy 1.x on MAC OS may not raise a singularity flag where Linux and Windows due. This issue appears to be addressed in numpy 2. Either way, this test should now be more in line with the goals of this test case (i.e., either (a) confirming that the matrix is singular and caught appropriately with nans in the DOP entries or (b) near singular even if the flag is not raised and the relevant DOP entries are non-physical).